// ADMIN-GUARDIAN/shared/schema.ts
export enum DirectTransferType {
  MINING_REWARD = 'mining_reward',
  TOKEN_LIQUIDITY = 'token_liquidity',
  NFT_TRANSFER = 'nft_transfer'
}